requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
clickhouse-driver==0.2.9
fastapi==0.115.4
h11==0.14.0
idna==3.10
kafka-python @ git+https://github.com/dpkp/kafka-python.git@a6d0579d3cadd3826dd364b01bc12a2173139abc
pika==1.3.2
pydantic==2.9.2
pydantic_core==2.23.4
pytz==2024.2
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.0
